% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_age.R
\name{CK}
\alias{CK}
\title{Coale and Kisker Method of Age Completion}
\usage{
CK(rates, type, ages, years, old_ages, m_end)
}
\arguments{
\item{rates}{mortality rates in a rectangular array with ages (on the rows)
and calendar year (on the columns). Can be a matrix or a vector}

\item{type}{specifies the type of rates supplied. Takes the following values:
"central" for central death rates, "prob" for 1-year death probabilities,
"force" for force of mortality}

\item{ages}{age vector for \code{rates}}

\item{years}{year vector for \code{rates}}

\item{old_ages}{old ages to be completed for}

\item{m_end}{a constant or vector specifying the central death rates at the final age}
}
\value{
Central death rates in a rectangular array for all ages and calendar years
}
\description{
Implements the Coale and Kisker method of age completion.
}
\examples{

}
