% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_age.R
\name{kannisto}
\alias{kannisto}
\title{Kannisto Method of Age Completion}
\usage{
kannisto(rates, type, ages, years, old_ages, fitted_ages)
}
\arguments{
\item{rates}{mortality rates in a rectangular array with ages (on the rows)
and calendar year (on the columns)}

\item{type}{specifies the type of rates supplied. Takes the following values:
"central" for central death rates, "prob" for 1-year death probabilities,
"force" for force of mortality}

\item{ages}{age vector for \code{rates}}

\item{years}{year vector for \code{rates}}

\item{old_ages}{old ages to be completed for}

\item{fitted_ages}{ages for which parameters are to be estimated from}
}
\value{
Force of mortality in a rectangular array for all ages and calendar years
}
\description{
Implements the Kannisto method of age completion
}
\examples{

}
