% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_stats.R
\name{plot_exp_cfl}
\alias{plot_exp_cfl}
\title{Plot Curtate Future Lifetime Forecasts}
\usage{
plot_exp_cfl(exp_cfl_hist, years_hist, exp_cfl_for, years_for, level = 95)
}
\arguments{
\item{exp_cfl_hist}{vector of expected curtate future lifetime for historical years}

\item{years_hist}{vector of historical years}

\item{exp_cfl_for}{matrix of expected curtate future lifetime for forecasted years
with simulation number (on the rows) and calendar year (on the columns)}

\item{years_for}{vector of forecasted years}

\item{level}{desired confidence level with 95\% as default}
}
\value{

}
\description{
Plots historical and forecasted expected curtate future lifetime.
}
\examples{

}
