% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_stats.R
\name{plot_surv_fn_for}
\alias{plot_surv_fn_for}
\title{Plot Survival Function Forecasts}
\usage{
plot_surv_fn_for(surv_fn_for, init_age, target_year, level = 95)
}
\arguments{
\item{surv_fn_for}{3D array of forecasted survival functions with survival time
(on the rows) and calendar year (on the columns) and simulation
number (3rd dimension)}

\item{init_age}{initial age for which \code{surv_fn_for} was calculated at}

\item{target_year}{year for which the survival function is plotted for}

\item{level}{desired confidence level with 95\% as default}
}
\value{
returns \code{NULL}
}
\description{
Plots forecasted survival functions for a desired cohort.
}
\examples{

}
