% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_stats.R
\name{psurviv}
\alias{psurviv}
\title{Distribution of Survival Function}
\usage{
psurviv(surviv_rates, surviv_time)
}
\arguments{
\item{surviv_rates}{vector of survival rates}

\item{surviv_time}{vector of survival times}
}
\value{
Distribution function
}
\description{
Distribution of the Survival Function
}
\details{
Linear interpolation is performed between the discrete points of the survival function
}
\examples{

}
