% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/price_of_risk.R
\name{q2survival}
\alias{q2survival}
\title{Convert death probabilities to survival rates}
\usage{
q2survival(qx)
}
\arguments{
\item{qx}{vector, matrix or 3D array of 1-year death probabilities with age
(on the rows) and calendar year (on the columns)}
}
\value{
survival rates in the same format as \code{qx}
}
\description{
Converts 1-year death probabilities to the associated survival function.
}
\examples{

}
