% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_stats.R
\name{surviv}
\alias{surviv}
\alias{psurviv}
\alias{qsurviv}
\title{The Survival Function}
\usage{
psurviv(surviv_fun, surviv_time)

qsurviv(surviv_fun, surviv_prob)
}
\arguments{
\item{surviv_fun}{vector of survival function}

\item{surviv_time}{vector of survival times}

\item{surviv_prob}{vector of survival probabilities}
}
\description{
Distribution and quantile function  of the survival function.
}
\details{
Linear interpolation is performed between the discrete points of the survival function
}
