% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/price_of_risk.R
\name{survival2q}
\alias{survival2q}
\title{Convert survival function to death probabilities}
\usage{
survival2q(surv, ages, years = NULL)
}
\arguments{
\item{surv}{vector, matrix or 3D array of the survival function with survival time starting from 0
(on the rows) and calendar year (on the columns) and simulation number (3rd dimension)}

\item{ages}{vector of desired ages for the resulting 1-year death probabilities}

\item{years}{optional vector of years for \code{surv}. If not supplied, then the column names
of \code{surv} will be preserved}
}
\value{
associated 1-year death probabilities in the same format as \code{surv}
}
\description{
Converts the survival function to the associated 1-year death probabilities.
}
\examples{

}
