% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/price_of_risk.R
\name{survivalP2Q}
\alias{survivalP2Q}
\title{Survival Function Transformation}
\usage{
survivalP2Q(StP, method, lambda, t = NULL, ages = NULL, years = NULL)
}
\arguments{
\item{StP}{vector, matrix or 3D array of the survival function under the P-measure with
survival time (on the rows) and calendar year (on the columns)}

\item{method}{character string representing the distortion risk measure to be used. See "Details".}

\item{lambda}{parameter associated with the distortion risk measure}

\item{t}{optional vector of survival time}

\item{ages}{optional vector of ages}

\item{years}{optional vector of years}
}
\value{
the transformed survival function under the Q-measure
}
\description{
Transforms the survival function from the real world P-measure to the risk-neutral
Q-measure according to the specified distortion risk measure
}
\details{
The distortion risk measures and their strings are as follows
\itemize{
\item "wang": Wang Transform
\item "ph": Proportional Hazard Transform
\item "dp": Dual-power Transform
\item "dadp": Denneberg's Absolute Deviation Principle
\item "exp": Exponential Transform
\item "log": Logarithmic Transform
}
}
\examples{

}
