% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_age.R
\name{CK}
\alias{CK}
\title{Coale and Kisker Method}
\usage{
CK(rates, ages, old_ages, type = "central", m_end = 1, years = NULL)
}
\arguments{
\item{rates}{matrix or vector of mortality rates with age (on the rows) and
calendar year (on the columns). Vector is equivalent to a matrix with a
single column}

\item{ages}{vector of ages for \code{rates}}

\item{old_ages}{vector of old ages for which \code{rates} is to be completed for}

\item{type}{character string representing the type of mortality rate
supplied. Takes the following values: "central" for central death rates,
"prob" for 1-year death probabilities, "force" for force of mortality}

\item{m_end}{constant or vector specifying the central death rates at the
final age for each calendar year. If supplied as a vector, vector length
and number of columns for \code{rates} must be equal}

\item{years}{optional vector of years for \code{rates}. If not supplied, then the
column names of \code{rates} will be preserved}
}
\value{
matrix of central death rates for all ages and calendar years
}
\description{
Implements the Coale and Kisker method of mortality rate completion for old
ages.
}
\examples{

}
