% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_age.R
\name{DG}
\alias{DG}
\title{Denuit and Goderniaux Method of Age Completion}
\usage{
DG(
  rates,
  type,
  ages,
  years,
  old_ages,
  closure_age = 130,
  min_fit_age = 75,
  smoothing = FALSE
)
}
\arguments{
\item{rates}{mortality rates in a rectangular array with ages (on the rows)
and calendar year (on the columns)}

\item{type}{specifies the type of rates supplied. Takes the following values:
"central" for central death rates, "prob" for 1-year death probabilities,
"force" for force of mortality}

\item{ages}{age vector for \code{rates}}

\item{years}{year vector for \code{rates}}

\item{old_ages}{old ages to be completed for}

\item{closure_age}{age for which closure constraint is applied}

\item{min_fit_age}{model is fitted starting from this age}

\item{smoothing}{logical value indicating if smoothing is to be applied}
}
\value{
1-year death probabilities in a rectangular array for all ages and calendar years
}
\description{
Implements the Denuit and Goderniaux method of age completion.
}
\examples{

}
